+++ BSEAV/app/atlas/src/channel_sat.h
@@ -73,7 +73,7 @@ public:
     ~CChannelSat(void);
     CChannel * createCopy(CChannel * pChannel);
 
-    virtual eRet tune(void * id, CConfig * pConfig, bool bWaitForLock, uint16_t index = ANY_INDEX);
+    virtual eRet tune(void * id, CConfig * pConfig, bool bWaitForLock, unsigned index = ANY_INDEX);
     virtual eRet unTune(CConfig * pResourceLibrary, bool bFullUnTune = false, bool bCheckInTuner = true);
     virtual eRet readXML(MXmlElement * xmlElemChannel);
     virtual void writeXML(MXmlElement * xmlElemChannel);
