+++ BSEAV/app/atlas/src/channel_vsb.cpp
@@ -234,7 +234,7 @@ eRet CChannelVsb::tune(
         void *    id,
         CConfig * pConfig,
         bool      bWaitForLock,
-        uint16_t  index
+        unsigned  index
         )
 {
     eRet                         ret    = eRet_Ok;
