+++ BSEAV/app/atlas/src/nxclient/bluetooth_nx.h
@@ -64,7 +64,7 @@ class CBluetoothNx : public CBluetooth
 public:
     CBluetoothNx(
             const char *     name,
-            const uint16_t   number,
+            const unsigned   number,
             CConfiguration * pCfg
             );
     ~CBluetoothNx(void);
