+++ BSEAV/app/atlas/src/tuner_sat.cpp
@@ -115,7 +115,7 @@ void CTunerSatScanData::dump()
 
 CTunerSat::CTunerSat(
         const char *     name,
-        const uint16_t   number,
+        const unsigned   number,
         CConfiguration * pCfg
         ) :
     CTuner(name, number, eBoardResource_frontendSds, pCfg),
@@ -392,9 +392,9 @@ void CTunerSat::saveScanData(CTunerScanData * pScanData)
 void CTunerSat::doScan()
 {
     eRet                       ret            = eRet_Ok;
-    uint16_t                   major          = 1;
-    uint16_t                   numFreqToScan  = 0;
-    uint16_t                   numFreqScanned = 0;
+    unsigned                   major          = 1;
+    unsigned                   numFreqToScan  = 0;
+    unsigned                   numFreqScanned = 0;
     CChannelSat                chSat(_pCfg, this); /* temp channel we'll use to do tuning during scan */
     CTunerScanNotificationData notifyData(this);   /* notification data for reporting scan start/stop/progress */
