+++ BSEAV/app/atlas/src/tuner_vsb.cpp
@@ -117,7 +117,7 @@ void CTunerVsbScanData::dump()
 
 CTunerVsb::CTunerVsb(
         const char *     name,
-        const uint16_t   number,
+        const unsigned   number,
         CConfiguration * pCfg
         ) :
     CTuner(name, number, eBoardResource_frontendVsb, pCfg)
@@ -265,9 +265,9 @@ void CTunerVsb::saveScanData(CTunerScanData * pScanData)
 void CTunerVsb::doScan()
 {
     eRet                       ret            = eRet_Ok;
-    uint16_t                   major          = 1;
-    uint16_t                   numFreqToScan  = 0;
-    uint16_t                   numFreqScanned = 0;
+    unsigned                   major          = 1;
+    unsigned                   numFreqToScan  = 0;
+    unsigned                   numFreqScanned = 0;
     CChannelVsb                chVsb(_pCfg, this); /* temp channel we'll use to do tuning during scan */
     CTunerScanNotificationData notifyData(this);   /* notification data for reporting scan start/stop/progress */
     uint32_t *                 pFreq = NULL;
